+++ tests/testthat/test-var_grid.R
@@ -98,6 +98,33 @@ test_that("var_grid only xvar", {
     expect_equal(res$var, c("v", "w", "x", "y", "z"))
 })
 
+test_that("var_grid with disjoint", {
+    d <- data.frame(v = FALSE,
+                    w = FALSE,
+                    x = TRUE,
+                    y = c(TRUE, FALSE),
+                    z = c(TRUE, FALSE))
+
+    res <- var_grid(d,
+                    xvars = everything(),
+                    yvars = everything())
+
+    expect_true(is_tibble(res))
+    expect_equal(nrow(res), 10)
+    expect_equal(colnames(res), c("xvar", "yvar"))
+    expect_equal(res$xvar, c("v", "v", "v", "v", "w", "w", "w", "x", "x", "y"))
+    expect_equal(res$yvar, c("w", "x", "y", "z", "x", "y", "z", "y", "z", "z"))
+
+    res <- var_grid(d,
+                    xvars = everything(),
+                    yvars = everything(),
+                    disjoint = c("vw", "vw", "xyz", "xyz", "xyz"))
+
+    expect_equal(nrow(res), 6)
+    expect_equal(colnames(res), c("xvar", "yvar"))
+    expect_equal(res$xvar, c("v", "v", "v", "w", "w", "w"))
+    expect_equal(res$yvar, c("x", "y", "z", "x", "y", "z"))
+})
 
 test_that("var_grid errors", {
     d <- data.frame(v = FALSE,
@@ -121,4 +148,11 @@ test_that("var_grid errors", {
                           xvars = everything(),
                           yvars = everything()),
                  "`x` must be a matrix or a data frame")
+
+    expect_error(var_grid(d, xvars = everything(), yvars = everything(),
+                          disjoint = list("x")),
+                 "`disjoint` must be a plain vector")
+    expect_error(var_grid(d, xvars = everything(), yvars = everything(),
+                          disjoint = "x"),
+                 "The length of `disjoint` must be 0 or must be equal to the number of columns in `x`.")
 })